* commit 'c77ba5cc9d7e0dd7eb8a0fac982300a9547449f4':
  Changes suggested in review.
  Generated windows and iOS projects now support 64-bit.
  Added windows 64-bit support.
  Fixed all 64-bit warnings on iOS.
  Updated iOS to build 64-bit binaries. Project still contains a number of integer truncation warnings.

+++ Source/CSBackend/Platform/iOS/Core/DialogueBox/DialogueBoxSystem.h
@@ -98,7 +98,7 @@ namespace CSBackend
             /// @param ID
             /// @param Result
             //------------------------------------------------------
-            void OnSystemConfirmDialogResult(u32 in_id, CSCore::DialogueBoxSystem::DialogueResult in_result);
+            void OnSystemConfirmDialogResult(s64 in_id, CSCore::DialogueBoxSystem::DialogueResult in_result);
             //----------------------------------------------------
 			/// Destructor.
             ///
